Split the TryGetValue part of MA0186 into a new rule MA0221
MissingNotNullWhenAttributeOnEqualsAnalyzer declared two descriptors under
the MA0186 id: one for Equals parameters missing [NotNullWhen(true)], one
for IDictionary<,>.TryGetValue value parameters missing [MaybeNullWhen(false)].
It was the only id in the repository with two descriptors.

The documentation generator de-duplicates descriptors by id, so the READMEs
and the five shipped editorconfigs only ever described MA0186 as the Equals
rule, and the TryGetValue behavior was invisible outside docs/Rules/MA0186.md.
Consumers could not find the rule by title, nor disable one half without the
other.

The TryGetValue rule is now MA0221, with its own analyzer, code fixer,
documentation page and tests. The two fixers share the attribute-rewriting
logic, and each now validates that its attribute type exists before
registering the fix instead of returning an unchanged document from the fix
action.

RuleHelpUri gains EachRuleIdHasASingleDescriptor so a future duplicate id
fails a test rather than silently disappearing from the generated lists.
