+++ docs/Rules/MA0186.md
@@ -5,15 +5,11 @@ Sources: [MissingNotNullWhenAttributeOnEqualsAnalyzer.cs](https://github.com/mez
 
 ## Description
 
-This rule reports missing nullable attributes on method parameters to improve nullable reference type analysis:
+When implementing `Equals(object?)` or `IEquatable<T>.Equals(T?)`, the parameter should be annotated with `[NotNullWhen(true)]` to inform the compiler that when the method returns `true`, the parameter is guaranteed to be non-null.
 
-1. When implementing `Equals(object?)` or `IEquatable<T>.Equals(T?)`, the parameter should be annotated with `[NotNullWhen(true)]` to inform the compiler that when the method returns `true`, the parameter is guaranteed to be non-null.
-
-2. When implementing `IDictionary<TKey, TValue>.TryGetValue`, the `value` parameter should be annotated with `[MaybeNullWhen(false)]` to inform the compiler that when the method returns `false`, the parameter may be null.
+A related rule, [MA0221](MA0221.md), reports missing `[MaybeNullWhen(false)]` on `IDictionary<TKey, TValue>.TryGetValue` implementations.
 
 ## Motivation
-
-### Equals methods
 
 Without the `[NotNullWhen(true)]` attribute, the compiler cannot track that a successful equality comparison means the parameter is non-null. This can lead to unnecessary null checks or null-forgiving operators in code that follows an equality check.
 
@@ -45,33 +41,7 @@ if (myInstance.Equals(someObj))
 }
 ````
 
-### TryGetValue methods
-
-Without the `[MaybeNullWhen(false)]` attribute, the compiler cannot track that when `TryGetValue` returns `false`, the out parameter may be null.
-
-````c#
-// ❌ Without the attribute
-public bool TryGetValue(TKey key, out TValue? value)
-{
-    // Implementation
-}
-
-// ✅ With the attribute
-public bool TryGetValue(TKey key, [MaybeNullWhen(false)] out TValue? value)
-{
-    // Implementation
-}
-
-// Usage
-if (dictionary.TryGetValue(key, out var value))
-{
-    value.ToString(); // Compiler knows value is not null
-}
-````
-
 ## How to fix violations
-
-### For Equals methods
 
 Add the `[NotNullWhen(true)]` attribute to the parameter:
 
@@ -88,23 +58,6 @@ public class MyType : IEquatable<MyType>
     public bool Equals([NotNullWhen(true)] MyType? other)
     {
         return other is not null && /* comparison logic */;
-    }
-}
-````
-
-### For TryGetValue methods
-
-Add the `[MaybeNullWhen(false)]` attribute to the value parameter:
-
-````c#
-using System.Diagnostics.CodeAnalysis;
-using System.Collections.Generic;
-
-public class MyDictionary : IDictionary<string, string?>
-{
-    public bool TryGetValue(string key, [MaybeNullWhen(false)] out string? value)
-    {
-        // Implementation
     }
 }
 ````
