+++ docs/Rules/MA0221.md
@@ -1,0 +1,65 @@
+# MA0221 - TryGetValue method should use \[MaybeNullWhen(false)\] on the value parameter
+<!-- sources -->
+Sources: [MissingMaybeNullWhenAttributeOnTryGetValueAnalyzer.cs](https://github.com/meziantou/Meziantou.Analyzer/blob/main/src/Meziantou.Analyzer/Rules/MissingMaybeNullWhenAttributeOnTryGetValueAnalyzer.cs), [MissingMaybeNullWhenAttributeOnTryGetValueFixer.cs](https://github.com/meziantou/Meziantou.Analyzer/blob/main/src/Meziantou.Analyzer.CodeFixers/Rules/MissingMaybeNullWhenAttributeOnTryGetValueFixer.cs)
+<!-- sources -->
+
+## Description
+
+When implementing `IDictionary<TKey, TValue>.TryGetValue`, the `value` parameter should be annotated with `[MaybeNullWhen(false)]` to inform the compiler that the parameter may be null only when the method returns `false`.
+
+A related rule, [MA0186](MA0186.md), reports missing `[NotNullWhen(true)]` on `Equals` parameters.
+
+## Motivation
+
+Without the `[MaybeNullWhen(false)]` attribute, the compiler cannot track that the out parameter is non-null when `TryGetValue` returns `true`.
+
+````c#
+// ❌ Without the attribute
+public bool TryGetValue(TKey key, out TValue? value)
+{
+    // Implementation
+}
+
+// Usage
+if (dictionary.TryGetValue(key, out var value))
+{
+    value.ToString(); // Warning CS8602
+}
+
+// ✅ With the attribute
+public bool TryGetValue(TKey key, [MaybeNullWhen(false)] out TValue? value)
+{
+    // Implementation
+}
+
+// Usage
+if (dictionary.TryGetValue(key, out var value))
+{
+    value.ToString(); // No warning
+}
+````
+
+## How to fix violations
+
+Add the `[MaybeNullWhen(false)]` attribute to the value parameter:
+
+````c#
+using System.Diagnostics.CodeAnalysis;
+using System.Collections.Generic;
+
+public class MyDictionary : IDictionary<string, string?>
+{
+    public bool TryGetValue(string key, [MaybeNullWhen(false)] out string? value)
+    {
+        // Implementation
+    }
+}
+````
+
+## Configuration
+
+This rule is disabled by default as it's a suggestion for improved nullable annotations. To enable it, add the following to your `.editorconfig` file:
+
+````editorconfig
+dotnet_diagnostic.MA0221.severity = suggestion
+````
