+++ src/Meziantou.Analyzer/RuleIdentifiers.cs
@@ -220,6 +220,7 @@ internal static class RuleIdentifiers
     public const string EmptyLanguageAttributeInXmlComment = "MA0218";
     public const string MissingLanguageAttributeInXmlComment = "MA0219";
     public const string InvalidRegexConfiguration = "MA0220";
+    public const string MissingMaybeNullWhenAttributeOnTryGetValue = "MA0221";
 
     public static string GetHelpUri(string identifier)
     {
