+++ tests/Meziantou.Analyzer.Test/Rules/MissingNotNullWhenAttributeOnEqualsAnalyzerTests.cs
@@ -1,6 +1,3 @@
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.Testing;
 using CodeFixTest = Meziantou.Analyzer.Test.Harness.CSharpCodeFixTest<
     Meziantou.Analyzer.Rules.MissingNotNullWhenAttributeOnEqualsAnalyzer,
     Meziantou.Analyzer.Rules.MissingNotNullWhenAttributeOnEqualsFixer>;
@@ -9,18 +6,10 @@ namespace Meziantou.Analyzer.Test.Rules;
 
 public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
 {
-    // The analyzer declares two descriptors with the same MA0186 id, so the markup cannot tell them apart
-    private static CodeFixTest CreateTest()
-    {
-        var test = new CodeFixTest();
-        test.MarkupOptions = MarkupOptions.UseFirstDescriptor;
-        return test;
-    }
-
     [Fact]
     public Task Equals_Object_WithoutAttribute_ShouldReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             class Sample
             {
@@ -39,7 +28,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_Object_WithoutAttribute_ShouldFix()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             using System.Diagnostics.CodeAnalysis;
 
@@ -73,7 +62,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_Object_WithWrongAttributeValue_ShouldFix()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             using System.Diagnostics.CodeAnalysis;
 
@@ -105,82 +94,9 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     }
 
     [Fact]
-    public Task TryGetValue_IDictionary_ExplicitWithoutAttribute_ShouldReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections.Generic;
-
-            class MyDictionary : IDictionary<string, string?>
-            {
-                bool IDictionary<string, string?>.TryGetValue(string key, out string? {|MA0186:value|})
-                {
-                    value = null;
-                    return false;
-                }
-
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public ICollection<string> Keys => throw new System.NotImplementedException();
-                public ICollection<string?> Values => throw new System.NotImplementedException();
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public IEnumerator<KeyValuePair<string, string?>> GetEnumerator() => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_IDictionary_ExplicitWithAttribute_ShouldNotReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections.Generic;
-            using System.Diagnostics.CodeAnalysis;
-
-            class MyDictionary : IDictionary<string, string?>
-            {
-                bool IDictionary<string, string?>.TryGetValue(string key, [MaybeNullWhen(false)] out string? value)
-                {
-                    value = null;
-                    return false;
-                }
-
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public ICollection<string> Keys => throw new System.NotImplementedException();
-                public ICollection<string?> Values => throw new System.NotImplementedException();
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public IEnumerator<KeyValuePair<string, string?>> GetEnumerator() => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
     public Task Equals_Object_WithAttribute_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             using System.Diagnostics.CodeAnalysis;
 
@@ -201,7 +117,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_IEquatable_WithoutAttribute_ShouldReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             using System;
 
@@ -227,7 +143,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_IEquatable_WithAttribute_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             using System;
             using System.Diagnostics.CodeAnalysis;
@@ -254,7 +170,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_NonNullableParameter_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             class Sample
             {
@@ -273,7 +189,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task NotEqualsMethod_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             class Sample
             {
@@ -290,7 +206,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task PrivateEquals_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             class Sample
             {
@@ -307,7 +223,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task StaticEquals_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             class Sample
             {
@@ -324,7 +240,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_WrongSignature_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             class Sample
             {
@@ -341,7 +257,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_IEquatable_BothMethodsWithoutAttribute_ShouldReportBothDiagnostics()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             using System;
 
@@ -367,7 +283,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_IEquatable_ValueType_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             using System;
 
@@ -393,7 +309,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_NullableDisabled_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             #nullable disable
             class Sample
@@ -413,7 +329,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_NullableEnabled_ShouldReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             #nullable enable
             class Sample
@@ -433,7 +349,7 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
     [Fact]
     public Task Equals_NullableEnabledThenDisabled_ShouldNotReportDiagnostic()
     {
-        var test = CreateTest();
+        var test = new CodeFixTest();
         test.TestCode = """
             #nullable enable
             class Sample1
@@ -455,372 +371,6 @@ public sealed class MissingNotNullWhenAttributeOnEqualsAnalyzerTests
                 }
 
                 public override int GetHashCode() => 0;
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_IDictionary_WithoutAttribute_ShouldReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections.Generic;
-
-            class MyDictionary : IDictionary<string, string?>
-            {
-                public bool TryGetValue(string key, out string? {|MA0186:value|})
-                {
-                    value = null;
-                    return false;
-                }
-
-                // Other IDictionary members...
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public ICollection<string> Keys => throw new System.NotImplementedException();
-                public ICollection<string?> Values => throw new System.NotImplementedException();
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public IEnumerator<KeyValuePair<string, string?>> GetEnumerator() => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_IDictionary_WithoutAttribute_ShouldFix()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections.Generic;
-            using System.Diagnostics.CodeAnalysis;
-
-            class MyDictionary : IDictionary<string, string?>
-            {
-                public bool TryGetValue(string key, out string? {|MA0186:value|})
-                {
-                    value = null;
-                    return false;
-                }
-
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public ICollection<string> Keys => throw new System.NotImplementedException();
-                public ICollection<string?> Values => throw new System.NotImplementedException();
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public IEnumerator<KeyValuePair<string, string?>> GetEnumerator() => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-        test.FixedCode = """
-            using System.Collections.Generic;
-            using System.Diagnostics.CodeAnalysis;
-
-            class MyDictionary : IDictionary<string, string?>
-            {
-                public bool TryGetValue(string key, [MaybeNullWhen(false)] out string? value)
-                {
-                    value = null;
-                    return false;
-                }
-
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public ICollection<string> Keys => throw new System.NotImplementedException();
-                public ICollection<string?> Values => throw new System.NotImplementedException();
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public IEnumerator<KeyValuePair<string, string?>> GetEnumerator() => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_IDictionary_WithAttribute_ShouldNotReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections.Generic;
-            using System.Diagnostics.CodeAnalysis;
-
-            class MyDictionary : IDictionary<string, string?>
-            {
-                public bool TryGetValue(string key, [MaybeNullWhen(false)] out string? value)
-                {
-                    value = null;
-                    return false;
-                }
-
-                // Other IDictionary members...
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public ICollection<string> Keys => throw new System.NotImplementedException();
-                public ICollection<string?> Values => throw new System.NotImplementedException();
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public IEnumerator<KeyValuePair<string, string?>> GetEnumerator() => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_IDictionary_Twice_BothInvalid_ShouldReportDiagnostics()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections;
-            using System.Collections.Generic;
-
-            class MyDictionary : IDictionary<string, string?>, IDictionary<int, string?>
-            {
-                public bool TryGetValue(string key, out string? {|MA0186:stringValue|})
-                {
-                    stringValue = null;
-                    return false;
-                }
-
-                public bool TryGetValue(int key, out string? {|MA0186:intValue|})
-                {
-                    intValue = null;
-                    return false;
-                }
-
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public string? this[int key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-
-                ICollection<string> IDictionary<string, string?>.Keys => throw new System.NotImplementedException();
-                ICollection<int> IDictionary<int, string?>.Keys => throw new System.NotImplementedException();
-                ICollection<string?> IDictionary<string, string?>.Values => throw new System.NotImplementedException();
-                ICollection<string?> IDictionary<int, string?>.Values => throw new System.NotImplementedException();
-
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(int key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public bool ContainsKey(int key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<int, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(int key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                IEnumerator<KeyValuePair<string, string?>> IEnumerable<KeyValuePair<string, string?>>.GetEnumerator() => throw new System.NotImplementedException();
-                IEnumerator<KeyValuePair<int, string?>> IEnumerable<KeyValuePair<int, string?>>.GetEnumerator() => throw new System.NotImplementedException();
-                IEnumerator IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_IDictionary_Twice_OneInvalid_ShouldReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections;
-            using System.Collections.Generic;
-            using System.Diagnostics.CodeAnalysis;
-
-            class MyDictionary : IDictionary<string, string?>, IDictionary<int, string?>
-            {
-                public bool TryGetValue(string key, out string? {|MA0186:stringValue|})
-                {
-                    stringValue = null;
-                    return false;
-                }
-
-                public bool TryGetValue(int key, [MaybeNullWhen(false)] out string? intValue)
-                {
-                    intValue = null;
-                    return false;
-                }
-
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public string? this[int key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-
-                ICollection<string> IDictionary<string, string?>.Keys => throw new System.NotImplementedException();
-                ICollection<int> IDictionary<int, string?>.Keys => throw new System.NotImplementedException();
-                ICollection<string?> IDictionary<string, string?>.Values => throw new System.NotImplementedException();
-                ICollection<string?> IDictionary<int, string?>.Values => throw new System.NotImplementedException();
-
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(int key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public bool ContainsKey(int key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<int, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(int key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                IEnumerator<KeyValuePair<string, string?>> IEnumerable<KeyValuePair<string, string?>>.GetEnumerator() => throw new System.NotImplementedException();
-                IEnumerator<KeyValuePair<int, string?>> IEnumerable<KeyValuePair<int, string?>>.GetEnumerator() => throw new System.NotImplementedException();
-                IEnumerator IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_IDictionary_Twice_NoneInvalid_ShouldNotReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections;
-            using System.Collections.Generic;
-            using System.Diagnostics.CodeAnalysis;
-
-            class MyDictionary : IDictionary<string, string?>, IDictionary<int, string?>
-            {
-                public bool TryGetValue(string key, [MaybeNullWhen(false)] out string? stringValue)
-                {
-                    stringValue = null;
-                    return false;
-                }
-
-                public bool TryGetValue(int key, [MaybeNullWhen(false)] out string? intValue)
-                {
-                    intValue = null;
-                    return false;
-                }
-
-                public string? this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public string? this[int key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-
-                ICollection<string> IDictionary<string, string?>.Keys => throw new System.NotImplementedException();
-                ICollection<int> IDictionary<int, string?>.Keys => throw new System.NotImplementedException();
-                ICollection<string?> IDictionary<string, string?>.Values => throw new System.NotImplementedException();
-                ICollection<string?> IDictionary<int, string?>.Values => throw new System.NotImplementedException();
-
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string? value) => throw new System.NotImplementedException();
-                public void Add(int key, string? value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public bool ContainsKey(int key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<int, string?>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(int key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string?> item) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<int, string?> item) => throw new System.NotImplementedException();
-                IEnumerator<KeyValuePair<string, string?>> IEnumerable<KeyValuePair<string, string?>>.GetEnumerator() => throw new System.NotImplementedException();
-                IEnumerator<KeyValuePair<int, string?>> IEnumerable<KeyValuePair<int, string?>>.GetEnumerator() => throw new System.NotImplementedException();
-                IEnumerator IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_NotIDictionary_ShouldNotReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            class MyClass
-            {
-                public bool TryGetValue(string key, out string? value)
-                {
-                    value = null;
-                    return false;
-                }
-            }
-            """;
-
-        return test.RunAsync();
-    }
-
-    [Fact]
-    public Task TryGetValue_NonNullableValue_ShouldNotReportDiagnostic()
-    {
-        var test = CreateTest();
-        test.TestCode = """
-            using System.Collections.Generic;
-
-            class MyDictionary : IDictionary<string, string>
-            {
-                public bool TryGetValue(string key, out string value)
-                {
-                    value = "";
-                    return false;
-                }
-
-                // Other IDictionary members...
-                public string this[string key] { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
-                public ICollection<string> Keys => throw new System.NotImplementedException();
-                public ICollection<string> Values => throw new System.NotImplementedException();
-                public int Count => throw new System.NotImplementedException();
-                public bool IsReadOnly => throw new System.NotImplementedException();
-                public void Add(string key, string value) => throw new System.NotImplementedException();
-                public void Add(KeyValuePair<string, string> item) => throw new System.NotImplementedException();
-                public void Clear() => throw new System.NotImplementedException();
-                public bool Contains(KeyValuePair<string, string> item) => throw new System.NotImplementedException();
-                public bool ContainsKey(string key) => throw new System.NotImplementedException();
-                public void CopyTo(KeyValuePair<string, string>[] array, int arrayIndex) => throw new System.NotImplementedException();
-                public IEnumerator<KeyValuePair<string, string>> GetEnumerator() => throw new System.NotImplementedException();
-                public bool Remove(string key) => throw new System.NotImplementedException();
-                public bool Remove(KeyValuePair<string, string> item) => throw new System.NotImplementedException();
-                System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => throw new System.NotImplementedException();
             }
             """;
